Create Module_6
This Module_6 folder contains everything from last week's Module_5 folder, retooled to work on my Windows machine in VS Code. Module_5 was developed in Replit and there were some inconsistencies that needed to be fixed to get it working for VS Code, most importantly the addition of a .env file that contains PostgreSQL info. This folder has NOT yet been reformatted as required by step 2 of the module 6 homework assignment.